deepsearch/requirements.txt
deepsearch-toolkit==0.40.0
fastapi
uvicorn[standard]
python-dotenv